Convert code to Go 1
Most of the changes are the result of running 'go fix'.  The rest are
changes to accessors for the os.FileInfo.  The private information has
been moved into a Sys() accessor, which is an interface{} to
(*syscall.Stat_t) which is the raw underlying stat struct.  This is
actually makes things a bit easier to ensure accuracy on.