sure: AttMap as struct
Instead of storing the attributes as a map[string]string with everything
converted to strings, store all of the attributes in a simple struct
with proper types for the fields, and do the hard work at encode/decode
time.  With this, store the sha1 hash as 20 bytes instead of in hex.

This reduces the memory footprint of large scans by about half, and
speeds up the operations a bit.  However, it makes the code a good bit
more fragile, since the encoder/decoder have to also have knowledge of
which attributes apply to which file types.  The types of the attributes
are now also system specific.  Future work will be to devise a way of
storing the attributes as a struct, but keeping the code organized that
does the marshal and unmarshal operations.

Jira: GOS-3