sure: Reflection based attributes
Instead of a single structure that tries to hold all attribute types,
use different structure types for the different file types.  Implement
the encoder, decoder, and comparison functions using reflection, instead
of being hardcoded.  This reduces some of the performance gain gotten by
using the structure, and there is room for improvement.  But, this
reduces the fragility of the attribute types.  Adding fields or even new
types here only requires changes to the structs themselves, and possibly
two tables if new types are added.

Jira: GOS-4