sure: Implement warnings for attribute changes
When decoding, if the attributes don't exactly match the structure,
print out a warning for fields that are extra or missing.  This helps
inform the user in case there are changes across versions of this
program.

Keep track of the warnings printed, however, so that we are only warning
once.