sure: Use just raw rdev value, not maj/min
Although the platform does have a notion of major/minor values to the
device node value, this is not well represented in the C library, and
not really relevant to userspace.

Replace the Devmaj and Devmin fields with just a Rdev field that
directly encodes the st_rdev field from the stat structure.  Although
this will produce warnings at first with the device nodes, it will make
this code easier to port to other platforms.