Port gosure to Darwin (MacOS)
Darwin is similar to Linux, but there are some minor differences:

- Only Linux has O_NOATIME
- Linking to openssl is tricky
- The 'sha1sum' program may be called 'gsha1sum' because of Homebrew
- The Ctim and Mtim Stat_t fields are called Ctimespec and Mtimespec

Use conditional build specs to support these changes, such that the
code will work on both Darwin and Linux.

Note that Apple no longer provides openssl.  I've hardcoded the
include and lib paths for homebrew's default tree, but this would have
to be changed to support another installation method.