store: Make default compressed if only base given
In the situation where a filename is given, but has no extension at all,
use compression instead of without compression.  This allows something
like:

  -f /path/to/store

to use files /path/to/store.dat.gz and the likes.  If no compression is
desired, the ".dat" extension should be given, e.g.:

  -f /path/to/store.dat

and it is still possible to give the extension with the .gz suffix to
indicate a name with compression desired:

  -f /path/to/store.dat.gz