sure: Rewrite decoder to be push based
Instead of the decoder reading from a reader, implement it as a struct
that can have strings pushed to it for each line.  This can still be
used to read from a reader, but will make things much easier when
reading from weave files, which is most easily done by pushing.