Allow scan with unreadable file
Fixes #3, a scan with a file that is unreadable will successfully log
this.  However, the hash isn't updated, and the encoder currently
panics.

Resolve this by just skipping the empty attribute, rather than panicing.
On restore, it is left uninitialized, which compares the same as what
the unreadable has reads as.