Remove the 'linuxdir' package entirely
Although there is a slight performance increase to how linuxdir does
things (it reads directories, sorts by inode, and then stats the files).
It only applies on some filesystems, and only with extremely large
directories (millions of files).

Overall, it really isn't worth the tradeoff of having to bring in cgo
just for this.

Instead, use a simpler implementation that just stats the names as they
are in directory order.